Split orderhistory component to separate components for activity and user
